able to seperate out the Chat Component from the App
