+++ src/components/canvas.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+function Canvas() {
+
+    return(
+        <>
+        
+        </>
+    )
+}
+
+export default Canvas
